Fix bintr pbase issue, enable RISCV_INBOUND_JUMPS_ONLY by default with icache
Pay attention to this commit, as it may be faulty in rare untested cases. This commit makes room
for 4 bytes of zeroed memory at the beginning of the execute segment, so that we can have unchecked
jumps in the CPU loop, and instead check bounds during explicit CPU jumps (which there are less of).