Initial RV128I support
RV128I is not fully specified yet, and there is missing opcode information.
There is also bound to be offset and shifting bugs, as suble things like stack pointer offsets
and shift values change when transitioning from 64- to 128-bit ISA. There is a 128-bit emulator
with an example program in the emulator/emu128 folder.