Skip 0-length ELF segments, don't serialize small rodata segments, validate initial execute page
The very clever libfuzzer managed to trick the exec segment serializer by
creating a segment that wrapped around and setting PC to 0 (on the other side).
This is only possible because the emulator supports calling functions in the
guest virtual machine by appending a helper exit function to the execute
segment. It is done by adding 12 bytes: a STOP instruction, a relative JMP -4,
and finally a zero (illegal) instruction. These instructions allow us to have
an exit function to return to when returning from the guest function call.
This function was, unfortunately, able to wrap around the address space by
12 bytes. The fix is very simple, and additionally, we now check the page
of the initial PC, which will prevent this class of bugs.