bintr: Add emulator configuration to binary translation hash
The hash used to contain the entire compiler command, however I believe it makes more
sense to only add settings that can cause incompatibility. Whether or not you compile
with O1 or O2 shouldn't make clients reject it bashed on hash. Instead, now it will
only be rejected if the emulator is built with different settings, eg. C-ext.