Trim the size of the Machine structure a bit
This commit removes the unused CPU id, slow-path cpu page cache, current_exception (libtcc-only), and places execute segments in a dynamic std::vector.
The slow-path CPU page cache was used by read_next_instruction(), however it can be placed on the stack for precise simulation and passed as an argument.